+++ src/theme/advice.ts
@@ -32,9 +32,10 @@ export default {
       'border-(--b24ui-border-color) bg-(--b24ui-background) text-(--b24ui-color)',
       'context-dark:bg-(--ui-color-base-6)'
     ].join(' '),
-    leading: 'me-1.5 ms-2',
+    leading: 'me-1.5 ms-2 font-(--ui-font-weight-medium) text-(--ui-color-design-plain-content-icon-secondary)',
     leadingIcon: 'shrink-0 size-[42px]',
     leadingAvatar: 'shrink-0',
+    leadingAvatarIcon: 'text-(--ui-color-design-plain-content) bg-(--ui-color-base-8)',
     leadingAvatarSize: 'lg'
   },
   variants: {